requirements.txt
websockets>=12.0
aiohttp>=3.9
python-dotenv>=1.0
aiohttp-socks>=0.8
fastapi>=0.110
uvicorn[standard]>=0.29
